feat: Create class PersistencePickle
.Persistence class is now abstract
.Furthermore, pickle related behaviour
was moved to PersistencePickle
